[ISSUE-210][REGEX_HISTORY] Regex usage statistics issue
Regex usage statistics are not saved to the file when changing the
regex patterns file.

There is one more use case: switching between the pattern storage
files. To avoid losing the data, the regex usage history should be also
stored in the file in such  case.

This patch applies the desired behavior.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>